Change lattice->setup(volume) argument to be a CoordinateVector everywhere
Former-commit-id: 6d59a7b6011bfb11580fd2b8e0e7882d7343c5bf

+++ libraries/plumbing/backend_vector/defs.h
@@ -32,7 +32,7 @@
 // Define random number generator
 namespace hila {
     inline double random() { return mersenne(); }
-    inline void seed_random(int seed) {
+    inline void seed_random(unsigned long seed) {
 
         /* First seed the generator */
         seed_mersenne(seed);
